chatlist.py: ignore message without an username (server message)
Ignore all messages received without an username, this type of messages
is often used in system notice and should be ignored. Especially, some
servers with mandatory OTR requirement send an error if a non-OTR message
is received, if we reply these automatic messages with an automatic response,
an infinite loop could start and exhaust all system resources.

Signed-off-by: Yifeng Li <tomli@tomli.me>