chatlist.py: handle SIGTERM signal
When using chatlist as a system daemon, a SIGTERM signal is used by
the daemon manager to stop a daemon, but it's not handled in chatlist
and causes chatlist to quit immediately without closing connection or
saving data.

This commit adds a signal handler to call sys.exit() if SIGTERM is
received, effectively converting a SIGTERM to a SystemExit exception.

Signed-off-by: Yifeng Li <tomli@tomli.me>